+++ Gemfile
@@ -61,3 +61,4 @@ gem 'pry-rails'
 gem 'devise'
 gem 'simple_calendar', '~> 2.0'
 gem 'active_hash'
+gem 'ransack'
