--- spec/requests/posts_request_spec.rb
@@ -1,5 +1,0 @@
-require 'rails_helper'
-
-RSpec.describe "Posts", type: :request do
-
-end
